+++ butano/include/btn_bg_tiles_ptr.h
@@ -68,7 +68,7 @@ public:
 
     [[nodiscard]] optional<span<tile>> vram();
 
-    [[nodiscard]] size_t hash() const
+    [[nodiscard]] unsigned hash() const
     {
         return make_hash(_handle);
     }
@@ -98,7 +98,7 @@ private:
 template<>
 struct hash<bg_tiles_ptr>
 {
-    [[nodiscard]] size_t operator()(const bg_tiles_ptr& value) const
+    [[nodiscard]] unsigned operator()(const bg_tiles_ptr& value) const
     {
         return make_hash(value.id());
     }
